CONJ-1 - Timeout behavior is erratic. The actual reason is bug in socket timeout implementation in the JVM (at least reproducible on Windows),  that manifests in following socket behavior - once timeout on a socket occurs, after this the socket is "hosed" - after it, any timeout > 0 will result to immediate SocketTimeoutException. To workaround, timeout is reimplemented again, with the Timer. The implementation is moved from Protocol to Statement impementation , where it belongs to.

+++ src/main/java/org/mariadb/jdbc/internal/common/QueryException.java
@@ -83,7 +83,7 @@ public class QueryException extends Exception {
      * @param sqlState  the sqlstate
      */
     public QueryException(final String message,
-                          final short errorCode,
+                          final int errorCode,
                           final String sqlState) {
         super(message);
         this.errorCode = errorCode;
